ml_api/requirements.txt
flask
scikit-learn
pandas
numpy
Pillow
tensorflow
gunicorn
